apps/research-thera/evals/extraction/pyproject.toml
[project]
name = "extraction-evals"
version = "0.1.0"
description = "DeepEval quality evaluations for Qwen contact feedback issue extraction"
requires-python = ">=3.12"
dependencies = [
    "deepeval>=2.5.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
env = [
    "DEEPEVAL_TELEMETRY_OPT_OUT=YES",
]
